Bugfixes, island-finding upgrade.
Trigger a loss when we don't have valid moves remaining. I think
this was the not-learning problem we had.

Now trigger loss on islands of 3,4,9 and 14, OR >2 single islands
OR a double + any number of singles.

Update env validation. Ongoing: fiddling with hyperparams, but
learning works again. Hopefully no constraint simplex error with
fast learning rate (solved by action masking?).