Simplify hyperparams.
Currently training is still unstable. Solved sometimes not other times.

Going straight to the date solution. Part of the instability problem,
I suspect is that, when it stalls out at 6 or 7, which it does often,
it has quickly stumbled on those first pieces and can't
backtrack or learn that it needs to start over.

Anyway, a big change is coming in the date, which may make this
type of stall less likely. It'll need to be more general.